+++ apps/site/api/serializers/barcoded_serializer.py
@@ -2,7 +2,7 @@ import os
 from rest_framework import serializers
 from django.conf import settings
 from django.forms.fields import FileField
-from localground.apps.site.models.abstract.media import save_file_to_disk, generate_absolute_path, generate_relative_path
+from localground.apps.lib.helpers import upload_helpers, generic
 from localground.apps.site import models
 from localground.apps.site.api import fields
 from localground.apps.site.api.serializers.base_serializer import BaseNamedSerializer
@@ -32,13 +32,13 @@ class ScanSerializerCreate(BaseNamedSerializer):
     def process_file(self, file, owner):
         #save to disk:
         model_name_plural = models.Scan.model_name_plural
-        file_name_new = save_file_to_disk(owner, model_name_plural, file)
+        file_name_new = upload_helpers.save_file_to_disk(owner, model_name_plural, file)
         file_name, ext = os.path.splitext(file_name_new)
         
         # create thumbnail:
         from PIL import Image
         thumbnail_name = '%s_thumb.png' % file_name
-        media_path = generate_absolute_path(owner, model_name_plural)
+        media_path = upload_helpers.generate_absolute_path(owner, model_name_plural)
         im = Image.open(media_path + '/' + file_name_new)
         im.thumbnail([500, 500], Image.ANTIALIAS)
         im.save('%s/%s' % (media_path, thumbnail_name))
@@ -49,27 +49,24 @@ class ScanSerializerCreate(BaseNamedSerializer):
             'file_name_new': file_name_new,
             'file_name_thumb': thumbnail_name,
             'content_type': ext.replace('.', ''),
-            'virtual_path': generate_relative_path(owner, model_name_plural)
+            'virtual_path': upload_helpers.generate_relative_path(owner, model_name_plural)
         }
         
     def create(self, validated_data):
-        from localground.apps.lib.helpers import generic
+        '''
+        Overriding the create method to handle file processing
+        '''
         owner = self.context.get('request').user
         f = self.initial_data.get('media_file')
-        #raise Exception(f)
-        if f:
-            # ensure filetype is valid:
-            ext = os.path.splitext(f.name)[1]
-            ext = ext.lower().replace('.', '')
-            if ext not in self.ext_whitelist:
-                raise exceptions.UnsupportedMediaType(f,
-                    '{0} is not a valid map image file type. Valid options are: {1}'.format(
-                        ext, self.ext_whitelist)
-                )
+        
+        # ensure filetype is valid:
+        upload_helpers.validate_file(f, ScanSerializerCreate.ext_whitelist)
+        
+        # save it to disk
         extras = self.process_file(f, owner)
         extras.update({
             'uuid': generic.generateID(),
-            'status': models.StatusCode.objects.get(id=3),
+            'status': models.StatusCode.objects.get(id=3), #Make writeable field in serializer?
             'upload_source': models.UploadSource.objects.get(id=1),
             'owner': owner,
             'last_updated_by': owner,
