+++ apps/site/models/abstract/media.py
@@ -1,98 +1,14 @@
 from django.contrib.gis.db import models
 from localground.apps.site.models.abstract.audit import BaseAudit
 from localground.apps.site.models.abstract.mixins import ProjectMixin
-
+from localground.apps.lib.helpers import upload_helpers
 from tagging_autocomplete.models import TagAutocompleteField
 import base64
 from django.conf import settings
 from django.contrib.contenttypes.fields import GenericRelation
 import os
 import stat
-
-'''
-Utility File Path Methods
-'''
-def encrypt_media_path(host, model_name_plural, path):
-        from django.http import HttpResponse
-        import time
-        timestamp = int(time.time())
-        path = path + '#' + str(timestamp)
-        return '%s://%s/profile/%s/%s/' % (
-            settings.PROTOCOL,
-            host,
-            model_name_plural.replace(' ', '-'),
-            base64.b64encode(path)
-        )
-
-def encrypt_url(virtual_path, file_name):
-    return encrypt_media_path(virtual_path + file_name)
-
-def get_absolute_path(virtual_path):
-    return settings.FILE_ROOT + virtual_path
-
-def absolute_virtual_path(virtual_path, file_name):
-    return encrypt_media_path(virtual_path + file_name)
-
-def generate_relative_path(owner, model_name_plural):
-    return '/%s/media/%s/%s/' % (
-        settings.USER_MEDIA_DIR,
-        owner.username,
-        model_name_plural
-    )
-
-def generate_absolute_path(owner, model_name_plural):
-    return '%s/media/%s/%s' % (
-        settings.USER_MEDIA_ROOT,
-        owner.username,
-        model_name_plural
-    )
-
-def make_directory(path):
-    '''
-    I had problems making os.makedirs(path) work in terms of
-    setting the appropriate permissions, so I'm using this looping
-    function instead.  FYI, the user account needs to be the apache
-    account.  Any other way ends in tears.
-    '''
-    from pwd import getpwnam
-    uid = os.getuid()
-    gid = getpwnam(settings.GROUP_ACCOUNT).pw_gid
-    # same as 775:
-    permissions = stat.S_IRWXU | stat.S_IRWXG | stat.S_IROTH | stat.S_IXOTH
-    if not os.path.exists(path):
-        p = ""
-        paths = path.split("/")
-        paths.reverse()
-        while len(paths) > 0:
-            p += paths.pop() + '/'
-            if not os.path.exists(p):
-                # print '"%s" does not exist' % p
-                os.mkdir(p)
-                os.chown(p, uid, gid)
-                os.chmod(p, permissions)
-
-def save_file_to_disk(owner, model_name_plural, file):
-    # create directory if it doesn't exist:
-    media_path = generate_absolute_path(owner, model_name_plural)
-    if not os.path.exists(media_path):
-        make_directory(media_path)
-
-    # derive file name:
-    file_name, ext = os.path.splitext(file.name.lower())
-    file_name = ''.join(
-        char for char in file_name if char.isalnum()).lower()
-    if os.path.exists(media_path + '/%s%s' % (file_name, ext)):
-        from time import time
-        seconds = int(time())
-        file_name = '%s_%s' % (file_name, seconds)
-    file_name_new = '%s%s' % (file_name, ext)
-
-    # write request file stream to disk:
-    destination = open(media_path + '/' + file_name_new, 'wb+')
-    for chunk in file.chunks():
-        destination.write(chunk)
-    destination.close()
-    return file_name_new
+from rest_framework import exceptions
 
 class BaseMedia(BaseAudit):
 
@@ -125,25 +41,25 @@ class BaseMedia(BaseAudit):
         app_label = 'site'
 
     def get_absolute_path(self):
-        return get_absolute_path(self.virtual_path)
+        return upload_helpers.get_absolute_path(self.virtual_path)
 
     def absolute_virtual_path(self):
-        return absolute_virtual_path(self.virtual_path, self.file_name_new)
+        return upload_helpers.absolute_virtual_path(self.virtual_path, self.file_name_new)
 
     def absolute_virtual_path_orig(self):
-        return absolute_virtual_path(self.virtual_path, self.file_name_orig)
+        return upload_helpers.absolute_virtual_path(self.virtual_path, self.file_name_orig)
 
     def save_upload(self, *args, **kwargs):
         raise NotImplementedError
 
     def generate_relative_path(self):
-        return generate_relative_path(self.owner, self.model_name_plural)
+        return upload_helpers.generate_relative_path(self.owner, self.model_name_plural)
 
     def generate_absolute_path(self):
-        return generate_absolute_path(self.owner, self.model_name_plural)
+        return upload_helpers.generate_absolute_path(self.owner, self.model_name_plural)
 
     def _encrypt_media_path(self, path, host=None):
-        return encrypt_media_path(self.host, self.model_name_plural, path)
+        return upload_helpers.encrypt_media_path(self.host, self.model_name_plural, path)
 
     def encrypt_url(self, file_name):
         # return self.virtual_path + file_name
@@ -151,7 +67,7 @@ class BaseMedia(BaseAudit):
 
     @classmethod
     def make_directory(cls, path):
-        make_directory(path)
+        upload_helpers.make_directory(path)
 
     def can_view(self, user, access_key=None):
         return self.project.can_view(user=user, access_key=access_key)
